requirements.txt
pyrogram>=2.0.106
yt-dlp>=2023.12.30
tgcrypto>=1.2.5
python-dotenv>=1.0.0
aiofiles>=23.2.1
aiohttp>=3.9.1
motor>=3.3.2
pymongo>=4.6.0
dnspython>=2.4.2
